requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2024.12.14
charset-normalizer==3.4.0
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==3.10
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.32.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.5.5
urllib3==2.2.3
wrapt==1.11.1
